+++ components/src/core/DropdownToolbar/DropdownToolbar.tsx
@@ -159,7 +159,7 @@ const DropdownToolbarRender: React.ForwardRefRenderFunction<unknown, DropdownToo
         },
         [onOpen]
     );
-    
+
     useEffect(() => {
         console.warn(
             `DropdwonToolbar component has been deprecated and will be removed in version 6.0.0. Instead of DropdwonToolbar use ToolbarMenu component in regular Toolbar like Mui Appbar, BLUI Appbar.`