+++ components/src/core/index.ts
@@ -12,3 +12,4 @@ export * from './UserMenu';
 export * from './DropdownToolbar';
 export * from './AppBar';
 export * from './ThreeLiner';
+export * from './ToolbarMenu';