+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v5.4.0 (Unreleased)
+## v5.4.0-beta.1 (Unreleased)
 
 ### Added